Fixed some bugs with the ID expansion, changed the intersection code.
Unfortunately, the IDs weren't properly being expanded when saved into the
indices; this has been corrected. Additionally, the code to compute the
intersection of several tag indices has been changed: Now a single table
is built from the smallest index, and all of the contents of the other
tables are looked up in this, with the first byte of the string tracking
the reference count. This has removed the overhead of building the other
tables, while providing the same short-circuit code to abort if there is
no intersection as existed in the previous implementation.