Add a ntx_append function, and clean up some internal API semantics.
The ntx_append function now replaces all usages of gzopen(.., "a"),
and thus removes all direct zlib interaction from all first level
functions except ntx_list, where it is used for iteration and loading.

Some other internal APIs were clarified to prevent false positives
and negatives for the errors, and ntx_tagstolist was changed to
accept an argument for the ID of an entry.