Many small update/output fixes, and updated documentation.
The output of the tag and list commands are now standardized without a
trailing newline, and the main index is now properly updated after a
note is altered. A new check for malloc failure was also added.

As well, some outdated comments were fixed/removed, and an old warning
about index updates was removed from the help output.