ntx/rb NTXServer cleanups, and addition of the NTXClient class.
NTXServer was cleaned up, to be more readable, fix several bugs, and to
store notes in a Hash by ID, instead of storing them less efficiently
in Sets.

NTXClient was added, as a thin wrapper around client -> server
communication. It exposes a method for each request which can be made
to the server, as well as a receive method to retrieve a Response
from the transport, and a handle_response method to register callbacks
as closures for each Response type.