Suppress error output from a missing index in ntx_list().
The E_FACCESS exception thrown by failing to open the index file in the
ntx_list function when no arguments are passed is now suppressed, as it
was misleading. This case commonly occurred when the ntx database was
empty, and has never occurred in any other scenario during integration
testing.